requirements.txt
pandas
requests
alpaca_trade_api
google-auth
google-api-python-client
google-auth-httplib2
google-auth-oauthlib